Import cleanup

+++ src/main/java/dev/hilligans/ourcraft/client/rendering/MeshHolder.java
@@ -2,22 +2,22 @@ package dev.hilligans.ourcraft.client.rendering;
 
 public class MeshHolder {
 
-    public int id = -1;
-    public int oldID = 0;
+    public long id = -1;
+    public long oldID = 0;
 
     public int index = 0;
     public int length = 0;
 
     public int meshTexture = 0;
 
-    public MeshHolder set(int id, int length) {
+    public MeshHolder set(long id, int length) {
         oldID = id;
         this.id = id;
         this.length = length;
         return this;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 }
